src/main/java/ua/lastbite/userservice/validation/NameValidator.java
package ua.lastbite.userservice.validation;

import jakarta.validation.ConstraintValidator;
import jakarta.validation.ConstraintValidatorContext;

public class NameValidator implements ConstraintValidator<ValidName, String> {

    private static final String NAME_PATTERN = "^[A-Za-z\\s'-]+$";

    @Override
    public boolean isValid(String name, ConstraintValidatorContext constraintValidatorContext) {
        if (name == null || name.trim().isEmpty()) {
            return true; // Let @NotBlank handle the case
        }

        return name.matches(NAME_PATTERN);
    }
}
